separate suites into mods and make mods 1:1 with files
